feat: Add demo video (Git LFS) and fix token revocation issues
- Add demo video (52 MB) using Git LFS for MCP Server authentication flow
- Fix RefreshTokenRepository return type (long -> int) for cleanup job
- Fix MCP server expired device code handling
- Improve error messages in frontend authorization page
- Update README with demo video section

+++ backend/src/main/java/com/searchlab/repository/RefreshTokenRepository.java
@@ -51,9 +51,9 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
 
     /**
      * Delete expired refresh tokens (for cleanup job).
-     * FIXED: Returns long (count of deleted rows), not a collection.
+     * FIXED: Returns int (count of deleted rows) - Spring Data JPA modifying queries can only return void or int/Integer.
      */
     @Modifying
     @Query("DELETE FROM RefreshToken r WHERE r.expiresAt < :cutoff")
-    long deleteByExpiresAtBefore(@Param("cutoff") LocalDateTime cutoff);
+    int deleteByExpiresAtBefore(@Param("cutoff") LocalDateTime cutoff);
 }
